Log in start view for default FxA config (supports #3389)
